irgen: rename to ir/irconv
We may have irgen package later that generates IR-related code.

The ir/irconv package does not generate anything, it converts
php-parser AST into our own IR.
