+++ src/ir/irconv/irconv.go
@@ -1,4 +1,4 @@
-package irgen
+package irconv
 
 import (
 	"fmt"
