+++ src/phpgrep/matcher_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	"github.com/VKCOM/noverify/src/ir"
+	"github.com/VKCOM/noverify/src/ir/irconv"
 	"github.com/VKCOM/noverify/src/ir/irutil"
-	"github.com/VKCOM/noverify/src/irgen"
 	"github.com/VKCOM/noverify/src/php/parseutil"
 )
 
@@ -16,7 +16,7 @@ func mustParse(t testing.TB, code string) ir.Node {
 	if err != nil {
 		t.Fatalf("parse `%s`: %v", code, err)
 	}
-	irnode := irgen.ConvertNode(n)
+	irnode := irconv.ConvertNode(n)
 	if n, ok := irnode.(*ir.ExpressionStmt); ok {
 		return n.Expr
 	}
